feat: Add initial UnitedStatesMapSVG component; move components into subfolders
