+++ src/components/UnitedStatesMapSVG/UnitedStatesMapSVG.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const UnitedStatesMapSVG = () => (
+  <svg
+    aria-label="A map of the United States, showing which states I have visited."
+  >
+  </svg>
+);
+
+export default UnitedStatesMapSVG;
